Updated the sprites code to include the new apples variables to load in the apples, depending on the type of tree it is and add it to the specific groups in the sprite class.
Then created a new def to handle when to create the apples (20% chance to create one per game rendering, then add them to the left and top corner of the tree.
When the tree is created it adds the generic "apple" to the tree itself using generic class for pos, surf, group and z to determine the layering.

For some strange reason, its not rendering automatically without the print statement
                "print(f"Creating apple at: ({x}, {y})")  # Debugging line"

if i try to comment it out - the apples dont render... Need to try and debug this