Updated the process to know how to refresh the soil with no water when the day is reset
in Soil py created new remove water, this fines the water sprites and kills them (Not the actual sprite as its been converted to a "W" sprite. If this is correct then it cleans up the physical grid by finding any cells with a "W" variable and then removing those, to leave a clean field slate

on the level py we had to adjust the reset to say remove the water in the soil layer which is the def from the soil python file.